Update MathUtilities.java
Replace finite number check with my utility version, so that this library can be used with Java 6 clients. Unfortunately, this is currently a real issue, but those clients will be moving to Java 7 soon. I'll have to check when Java added the finite number check to the Float and Double classes.